basic tester code for the balancing robot
I think this version should be usable to control the motors in response to the IMU's input.  All bluetooth controls have been removed so it's just going off the IMU data.  One thing that still remains to be queried: does the PID loop still think it has to compensate for a full 2D plane or does it think it's in 1D?  Best way to find out is to build the robot and test.